Minor parsing of posts, fixed URL issue
Added protocol to the URLs so that links worked.
Moved link creation into a mixin.
Now read the title of the post, if it as J(unio)?r/i or S(enio)?r/i, we
add a class to make those stand out. Ultimate goal is to have 2 factors
(Interest and Certainty) to sort things in a similar fashion.